Add files via upload
This is a python project which is actually a game based project on perfect guess. This project aims in which the user has to guess the number accurate to the number choosen by computer. And it also returns the number of guesses made by the user to guess the perfect number.